+++ CHANGELOG.md
@@ -6,6 +6,9 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.x-dev
+
+- Add `--no-coverage` parameter option to skip code coverage generation.
 
 ## [3.2.1] - 2018-03-21